Added python-virtualenv to the dev tools.
From here, the rest of the python packages can be installed by the developer, in
their own virtual environment.

The package is in EPEL, not the main CentOS repos. We probably should be either
finding a way to make sure the repo is enabled, or just assume it is, and
install from there, but for now we're doing this.